Converts everything from Float32 to Float16

+++ NeuralObscura/Protocols.swift
@@ -9,8 +9,7 @@
 import Foundation
 import MetalPerformanceShaders
 
-//let textureFormat = MPSImageFeatureChannelFormat.float16
-let textureFormat = MPSImageFeatureChannelFormat.float32
+let textureFormat = MPSImageFeatureChannelFormat.float16
 
 enum CommandEncoderError: Error {
     case chainMisconfiguration
